change directory structure
