+++ lib/core/errors/exceptions.dart
