+++ lib/core/usecases/usecase.dart
@@ -1,7 +1,7 @@
 import 'package:dartz/dartz.dart';
 import 'package:equatable/equatable.dart';
 
-import '../error/failures.dart';
+import '../errors/failures.dart';
 
 abstract class UseCase<Success, Parameters> {
   Future<Either<Failure, Success>> call(Parameters parameters);
