+++ lib/features/player/data/models/clan_detail_model.dart
