+++ lib/features/player/data/models/clan_model.dart
