+++ lib/features/player/domain/repositories/player_repository.dart
@@ -1,0 +1,11 @@
+import 'package:dartz/dartz.dart';
+
+import '../../../../core/errors/failures.dart';
+import '../entities/player.dart';
+import '../entities/player_info.dart';
+
+abstract class PlayerRepository {
+  Future<Either<Failure, List<Player>>> fetchPlayerList(String search);
+
+  Future<Either<Failure, PlayerInfo>> fetchPlayerInfo(int accountId);
+}
