+++ lib/features/player/presentation/widgets/navigation_bar.dart
